added handling for disallowing organization delete if user exist

+++ lab2/src/main/java/com/cmpe275/lab2/service/impl/OrganizationServiceImpl.java
@@ -1,10 +1,12 @@
 package com.cmpe275.lab2.service.impl;
 
+import org.hibernate.HibernateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.cmpe275.lab2.dao.OrganizationDAO;
+import com.cmpe275.lab2.dao.PersonDAO;
 import com.cmpe275.lab2.model.Organization;
 import com.cmpe275.lab2.service.OrganizationService;
 
@@ -15,6 +17,9 @@ public class OrganizationServiceImpl implements OrganizationService {
 	
 	@Autowired
 	OrganizationDAO organizationDAO;
+	
+	@Autowired
+	PersonDAO personDAO;
 	
 	public Organization create(Organization organization) {
 		return organizationDAO.create(organization);
@@ -28,11 +33,14 @@ public class OrganizationServiceImpl implements OrganizationService {
 		return organizationDAO.update(organization);
 	}
 
-	public Organization delete(long id) {
+	public Organization delete(long id) throws HibernateException {
 		Organization organization = organizationDAO.read(id);
 		if(organization==null)
 			return null;
-		return organizationDAO.delete(organization);
+		if(personDAO.getPersonByOrganization(organization)!=null)
+			throw new HibernateException("Cannot delete organization");
+		else
+			return organizationDAO.delete(organization);
 	}
-
+	
 }
